Add Letters to HTML file script
PHP script that takes two arguments. The input file and the name for an output
file. Then converts the input text file into an html file. Specifically it
expects the input file to contain multiple letters prefixed with the phrase
"Editor, Smithfield Times" and for a new line character to appear at the end
of each paragraph.

 Changes to be committed:
	new file:   lettersToHTML.php